+++ natural_selection/__init__.py
@@ -7,13 +7,13 @@ import random
 import logging
 from datetime import datetime
 import warnings as w
-from typing import Callable, Any, Iterable, List, Union
+from typing import Callable, Any, Iterable, List, Union, Type
 import pickle
 
 import numpy as np
 
 from natural_selection.genetic_algorithms import Gene, Chromosome, Individual
-from natural_selection.genetic_programs import Node, GeneticProgram
+from natural_selection.genetic_programs import Node, GeneticProgram, random_generate
 
 from natural_selection.genetic_algorithms.operators.initialisation import initialise_population_random, alien_spawn_default
 from natural_selection.genetic_algorithms.operators.selection import selection_elites_top_n, selection_parents_two, selection_survivors_all
@@ -24,7 +24,7 @@ from natural_selection.genetic_algorithms.utils import clone_classic, default_sa
 
 from natural_selection.genetic_programs.operators.initialisation import initialise_population_full_method
 from natural_selection.genetic_programs.utils import GeneticProgramError
-import natural_selection.genetic_programs.functions as gp_func
+import natural_selection.genetic_programs.node_operators as gp_func
 
 
 class Island:
@@ -313,6 +313,59 @@ class Island:
         """
         return Chromosome(genes=genes, gene_verify_func=gene_verify_func, chromosome_properties=chromosome_properties)
 
+    def create_genetic_program(self,
+                               fitness_function: Callable = None,
+                               node_tree: Node = None,
+                               operators: List[Union[Type, gp_func.Operator]] = None,
+                               terminals: List[Union[str, int, float]] = None,
+                               max_depth: int = 3,
+                               min_depth: int = 1,
+                               growth_mode: str = 'grow',
+                               terminal_prob: float = 0.5,
+                               tree_generator: Callable = random_generate,
+                               name: str = None,
+                               species_type: str = None,
+                               add_to_population: bool = False,
+                               program_properties: dict = None
+                               ):
+        """
+        Wrapping function to create a new GeneticProgram. Useful when writing new initialisation functions. See GeneticProgram class.
+
+        Args:
+            fitness_function (Callable): Function with ``func(Node, island, **params)`` signature (default = None).
+            node_tree (Node): A starting node tree (default = None).
+            operators (list): List of all operators that nodes can be constructed from (default = None).
+            terminals (list): List of all terminals that can be included in the node tree, can be numeric or strings for variables (default = None).
+            max_depth (int): Maximum depth that node tree can grow (default = 3).
+            min_depth (int): Minimum depth that node tree must be (default = 1).
+            growth_mode (str): Type of tree growth method to use, "full" or "grow" (default = "grow").
+            terminal_prob (float): Probability of a generated node is a terminal (default = 0.5).
+            tree_generator (Callable): Function with to create the tree.
+            name (str): Name for keeping track of lineage (default = None).
+            species_type (str) : A unique string to identify the species type, for preventing cross polluting (default = None).
+            add_to_population (bool): Add this new program to the population (default = False).
+            program_properties (dict): For fitness functions, extra params may be given (default = None).
+
+        Returns:
+            program: Newly created GeneticProgram.
+        """
+        gp = GeneticProgram(fitness_function=fitness_function,
+                              node_tree=node_tree,
+                              operators=operators,
+                              terminals=terminals,
+                              max_depth=max_depth,
+                              min_depth=min_depth,
+                              growth_mode=growth_mode,
+                              tree_generator=tree_generator,
+                              name=name,
+                              species_type=species_type,
+                              program_properties=program_properties
+                              )
+        if add_to_population:
+            self.population.append(gp)
+            return gp
+
+
     def create_individual(self,
                           fitness_function : Callable,
                           name : str = None,
@@ -377,6 +430,7 @@ class Island:
         self.verbose_logging("init: complete")
 
         for p in self.population:
+            p.history.append({"init_island": self.name})
             self.__add_to_lineage(p)
 
     def __add_to_lineage(self, individual : Individual):
@@ -425,6 +479,7 @@ class Island:
 
         for i in migrants_for_adding:
             self.verbose_logging(f"migration: add {str(i)}")
+            i.history.append({"migration_island": self.name})
             self.population.append(i)
             self.unique_genome.append(i.unique_genetic_code())
             self.__add_to_lineage(i)
@@ -749,6 +804,7 @@ class Island:
                     # Else, add it effectively allowing "twins" to exist
                     alien_fitnesses.append(alien.fitness)
                     self.verbose_logging(f"evolve: add {str(alien)}")
+                    alien.history.append({"alien_init_island": self.name})
                     self.population.append(alien)
                     self.unique_genome.append(alien.unique_genetic_code())
                     self.__add_to_lineage(alien)